refactor: ui 패키지 이동 (#7)
