+++ app/src/main/java/com/example/ishopping/ui/cart/CartFragment.kt
@@ -1,4 +1,4 @@
-package com.example.ishopping.ui
+package com.example.ishopping.ui.cart
 
 import androidx.fragment.app.Fragment
 import com.example.ishopping.R
