+++ app/src/main/java/com/example/ishopping/ui/like/LikeFragment.kt
@@ -1,4 +1,4 @@
-package com.example.ishopping.ui
+package com.example.ishopping.ui.like
 
 import androidx.fragment.app.Fragment
 import com.example.ishopping.R
